New upstream version 8.5.0+dfsg

+++ ports/ci/travis/install_deps.sh
@@ -1,27 +1,55 @@
 #!/bin/bash
 
+# Webcamoid, webcam capture application.
+# Copyright (C) 2017  Gonzalo Exequiel Pedone
+#
+# Webcamoid is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Webcamoid is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Webcamoid. If not, see <http://www.gnu.org/licenses/>.
+#
+# Web-Site: http://webcamoid.github.io/
+
 if [ "${TRAVIS_OS_NAME}" = linux ] && [ "${ANDROID_BUILD}" != 1 ]; then
     mkdir -p .local/bin
+    qtIFW=QtInstallerFramework-linux-x64.run
 
     # Install Qt Installer Framework
-    wget -c http://download.qt.io/official_releases/qt-installer-framework/${QTIFWVER}/QtInstallerFramework-linux-x64.run
-    chmod +x QtInstallerFramework-linux-x64.run
-    export QT_QPA_PLATFORM=minimal
+    wget -c http://download.qt.io/official_releases/qt-installer-framework/${QTIFWVER}/${qtIFW} || true
 
-    ./QtInstallerFramework-linux-x64.run \
-        --script "$PWD/ports/ci/travis/qtifw_non_interactive_install.qs" \
-        --no-force-installations
+    if [ -e ${qtIFW} ]; then
+        chmod +x ${qtIFW}
 
-    cp -vf ~/Qt/QtIFW-${QTIFWVER/-*/}/bin/* .local/bin/
+        QT_QPA_PLATFORM=minimal \
+        ./QtInstallerFramework-linux-x64.run \
+            -v \
+            --script "$PWD/ports/ci/travis/qtifw_non_interactive_install.qs" \
+            --no-force-installations
+
+        cp -vf ~/Qt/QtIFW-${QTIFWVER/-*/}/bin/* .local/bin/
+    fi
+
+    appimage=appimagetool-x86_64.AppImage
 
     # Install AppImageTool
-    wget -c -O .local/bin/appimagetool-x86_64.AppImage https://github.com/AppImage/AppImageKit/releases/download/9/appimagetool-x86_64.AppImage
-    chmod +x .local/bin/appimagetool-x86_64.AppImage
+    wget -c -O .local/${appimage} https://github.com/AppImage/AppImageKit/releases/download/${APPIMAGEVER}/${appimage} || true
 
-    cd .local/bin
-    ./appimagetool-x86_64.AppImage --appimage-extract
-    cp -vf squashfs-root/usr/bin/* .
-    cd ../..
+    if [ -e .local/${appimage} ]; then
+        chmod +x .local/${appimage}
+
+        cd .local
+        ./${appimage} --appimage-extract
+        cp -rvf squashfs-root/usr/* .
+        cd ..
+    fi
 
     # Set default Docker command
     EXEC="docker exec ${DOCKERSYS}"
@@ -38,22 +66,20 @@ if [ "${ANDROID_BUILD}" = 1 ]; then
     unzip -q android-ndk-${NDKVER}-linux-x86_64.zip
 
     # Install Qt for Android
-    wget -c https://download.qt.io/archive/qt/${QTVER:0:3}/${QTVER}/qt-opensource-linux-x64-android-${QTVER}.run
-    chmod +x qt-opensource-linux-x64-android-${QTVER}.run
-    export QT_QPA_PLATFORM=minimal
+    wget -c https://download.qt.io/archive/qt/${QTVER:0:4}/${QTVER}/qt-opensource-linux-x64-${QTVER}.run
+    chmod +x qt-opensource-linux-x64-${QTVER}.run
 
-    ./qt-opensource-linux-x64-android-${QTVER}.run \
+    QT_QPA_PLATFORM=minimal \
+    ./qt-opensource-linux-x64-${QTVER}.run \
+        -v \
         --script "$PWD/../ports/ci/travis/qt_non_interactive_install.qs" \
         --no-force-installations
 elif [ "${DOCKERSYS}" = debian ]; then
     ${EXEC} apt-get -y update
 
-    if [ "${DOCKERIMG}" = ubuntu:trusty ]; then
+    if [ "${DOCKERIMG}" = ubuntu:xenial ]; then
         ${EXEC} apt-get -y install software-properties-common
-        ${EXEC} add-apt-repository ppa:beineri/opt-qt58-trusty
-    elif [ "${DOCKERIMG}" = ubuntu:xenial ]; then
-        ${EXEC} apt-get -y install software-properties-common
-        ${EXEC} add-apt-repository ppa:beineri/opt-qt58-xenial
+        ${EXEC} add-apt-repository ppa:beineri/opt-qt-${QTVER}-xenial
     fi
 
     ${EXEC} apt-get -y update
@@ -61,6 +87,7 @@ elif [ "${DOCKERSYS}" = debian ]; then
 
     # Install dev tools
     ${EXEC} apt-get -y install \
+        git \
         xvfb \
         g++ \
         clang \
@@ -76,16 +103,15 @@ elif [ "${DOCKERSYS}" = debian ]; then
         libgstreamer-plugins-base1.0-dev
 
     # Install Qt dev
-    if [ "${DOCKERIMG}" = ubuntu:trusty ] || \
-         [ "${DOCKERIMG}" = ubuntu:xenial ]; then
+    if [ "${DOCKERIMG}" = ubuntu:xenial ]; then
         ${EXEC} apt-get -y install \
-            qt58tools \
-            qt58declarative \
-            qt58multimedia \
-            qt58svg \
-            qt58quickcontrols \
-            qt58quickcontrols2 \
-            qt58graphicaleffects
+            qt${PPAQTVER}tools \
+            qt${PPAQTVER}declarative \
+            qt${PPAQTVER}multimedia \
+            qt${PPAQTVER}svg \
+            qt${PPAQTVER}quickcontrols \
+            qt${PPAQTVER}quickcontrols2 \
+            qt${PPAQTVER}graphicaleffects
     else
         ${EXEC} apt-get -y install \
             qt5-qmake \
@@ -93,13 +119,16 @@ elif [ "${DOCKERSYS}" = debian ]; then
             qtmultimedia5-dev \
             libqt5opengl5-dev \
             libqt5svg5-dev \
+            qtquickcontrols2-5-dev \
             qml-module-qt-labs-folderlistmodel \
             qml-module-qt-labs-settings \
             qml-module-qtqml-models2 \
             qml-module-qtquick-controls \
+            qml-module-qtquick-controls2 \
             qml-module-qtquick-dialogs \
             qml-module-qtquick-extras \
-            qml-module-qtquick-privatewidgets
+            qml-module-qtquick-privatewidgets \
+            qml-module-qtquick-templates2
     fi
 
     # Install FFmpeg dev
@@ -109,18 +138,15 @@ elif [ "${DOCKERSYS}" = debian ]; then
         libavformat-dev \
         libavutil-dev \
         libavresample-dev \
-        libswscale-dev
-
-    if [ "${DOCKERIMG}" != ubuntu:trusty ]; then
-        ${EXEC} apt-get -y install \
-            libswresample-dev
-    fi
+        libswscale-dev \
+        libswresample-dev
 elif [ "${DOCKERSYS}" = fedora ]; then
     ${EXEC} dnf install -y https://download1.rpmfusion.org/free/fedora/rpmfusion-free-release-${FEDORAVER}.noarch.rpm
     ${EXEC} dnf install -y https://download1.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-${FEDORAVER}.noarch.rpm
     ${EXEC} dnf -y update
 
     ${EXEC} dnf -y install \
+        git \
         file \
         which \
         xorg-x11-xauth \
@@ -134,6 +160,7 @@ elif [ "${DOCKERSYS}" = fedora ]; then
         qt5-qtmultimedia-devel \
         qt5-qtsvg-devel \
         qt5-qtquickcontrols \
+        qt5-qtquickcontrols2-devel \
         qt5-qtgraphicaleffects \
         ffmpeg-devel \
         gstreamer1-plugins-base-devel \
@@ -142,9 +169,10 @@ elif [ "${DOCKERSYS}" = fedora ]; then
         pulseaudio-libs-devel \
         jack-audio-connection-kit-devel
 elif [ "${DOCKERSYS}" = opensuse ]; then
-    ${EXEC} zypper -n update
+    ${EXEC} zypper -n dup
 
     ${EXEC} zypper -n in \
+        git \
         which \
         xauth \
         xvfb-run \
@@ -157,6 +185,8 @@ elif [ "${DOCKERSYS}" = opensuse ]; then
         libqt5-qtmultimedia-devel \
         libqt5-qtsvg-devel \
         libqt5-qtquickcontrols \
+        libqt5-qtquickcontrols2 \
+        libQt5QuickControls2-devel \
         libqt5-qtgraphicaleffects \
         ffmpeg-devel \
         gstreamer-plugins-base-devel \
@@ -178,19 +208,20 @@ elif [ "${TRAVIS_OS_NAME}" = osx ]; then
         jack \
         libuvc
 
+    qtIFW=QtInstallerFramework-mac-x64.dmg
+
     # Install Qt Installer Framework
-    wget -c http://download.qt.io/official_releases/qt-installer-framework/${QTIFWVER}/QtInstallerFramework-mac-x64.dmg
-    7z x -oqtifw QtInstallerFramework-mac-x64.dmg
-    7z x -oqtifw qtifw/5.hfsx
-    chmod +x qtifw/QtInstallerFramework-mac-x64/QtInstallerFramework-mac-x64.app/Contents/MacOS/QtInstallerFramework-mac-x64
+    wget -c http://download.qt.io/official_releases/qt-installer-framework/${QTIFWVER}/${qtIFW} || true
 
-    qtifw/QtInstallerFramework-mac-x64/QtInstallerFramework-mac-x64.app/Contents/MacOS/QtInstallerFramework-mac-x64 \
-        --script "$PWD/ports/ci/travis/qtifw_non_interactive_install.qs" \
-        --no-force-installations
+    if [ -e ${qtIFW} ]; then
+        hdiutil convert ${qtIFW} -format UDZO -o qtifw
+        7z x -oqtifw qtifw.dmg -bb
+        7z x -oqtifw qtifw/5.hfsx -bb
+        chmod +x qtifw/QtInstallerFramework-mac-x64/QtInstallerFramework-mac-x64.app/Contents/MacOS/QtInstallerFramework-mac-x64
 
-    # Install Syphon framework
-    wget -c https://github.com/Syphon/Simple/releases/download/version-3/Syphon.Simple.Apps.3.zip
-    unzip Syphon.Simple.Apps.3.zip
-    mkdir -p Syphon
-    cp -Rvf "Syphon Simple Apps/Simple Client.app/Contents/Frameworks/Syphon.framework" ./Syphon
+        qtifw/QtInstallerFramework-mac-x64/QtInstallerFramework-mac-x64.app/Contents/MacOS/QtInstallerFramework-mac-x64 \
+            -v \
+            --script "$PWD/ports/ci/travis/qtifw_non_interactive_install.qs" \
+            --no-force-installations
+    fi
 fi
